third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/forms/datalist/datalist-inside-shadow-dom-expected.txt

CONSOLE ERROR: line 16: Uncaught TypeError: host.createShadowRoot is not a function
